+++ tests/Fixture/ExampleEnum.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Tests\Torr\SimpleNormalizer\Fixture;
+
+enum ExampleEnum
+{
+	case Test;
+}